+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+# 3.3.0 / 2018-12-13
+
+- Add basic secrets support and docs
+
+# 3.2.0 / 2018-12-10
+
+Additions:
+
+- Added support for basic http auth
+
+Changes:
+
+- Moved server docs to main repo
+
+Fixes:
+
+- Fixed rateLimit field wrong format in config docs
+
 # 3.1.1 / 2018-09-19
 
 - Update npm deploy token for travis